Tests and fixes of the LoggerBase classes
* Implementation of the ConsoleLogger class
* Making LoggerBase a IEventRouter
* Tests of the catching of messages (the format of the message is still
an issue)
* Removal and refactoring of obsolete or unused classes

+++ src/deveeldb/Deveel.Data.Sql.Objects/SqlYearToMonth.cs
@@ -77,7 +77,7 @@ namespace Deveel.Data.Sql.Objects {
 				if (months == null)
 					throw new NullReferenceException();
 
-				return (months.Value/12);
+				return ((double)months.Value/12);
 			}
 		}
